combinations_with_replacement: change to allow lazy version to go beyond int upper bounds

+++ comb/comb.v
@@ -88,7 +88,7 @@ pub fn (mut o CombinationsIter) next() ?[]f64 {
 // Using f64 array instead of generic while waiting on https://github.com/vlang/v/issues/7753
 pub fn combinations_with_replacement(data []f64, r int) [][]f64 {
 	mut iter := new_combinations_with_replacement_iter(data, r)
-	mut result := [][]f64{cap: iter.size}
+	mut result := [][]f64{cap: int(iter.size)}
 	for _ in 0 .. iter.size {
 		if comb := iter.next() {
 			result << comb
@@ -99,11 +99,11 @@ pub fn combinations_with_replacement(data []f64, r int) [][]f64 {
 
 pub struct CombinationsWithReplacementIter {
 mut:
-	pos    int
+	pos    u64
 	idxs   []int
 pub:
 	repeat int
-	size   int
+	size   u64
 	data   []f64
 }
 
@@ -117,7 +117,7 @@ pub fn new_combinations_with_replacement_iter(data []f64, r int) CombinationsWit
 			repeat: r
 		}
 	}
-	size := int(fun.n_combos_w_replacement(n, r))
+	size := fun.n_combos_w_replacement(n, r)
 	idxs := []int{len: r, init: 0}
 	return CombinationsWithReplacementIter{
 		data: data
